+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# main
+- Implement port enumeration for Windows.
+
 # Version 0.1.0-alpha2 - 2021-11-07
 - Add `SerialPort::available_ports()`, which for now only works on Linux.